// src/modules/user/types/user.enums.ts
export enum PermissionEnum {
  LANDLORD = "landlord",
  TENANT = "tenant",
  SUPERVISOR = "supervisor",
  RENTRIX_REP = "representative",
  PROPERTY_ID = "property-id",
  COMMUNICATION_LOGS = "communication-logs",
  ESCROW_DISPUTE_LOGS = "escrow-dispute-logs",
  AI_ML_LOGS = "ai-ml-logs",
  ADMIN_ACTION_LOGS = "admin-action-logs",
  SYS_MAINTENANCE_LOGS = "system-maintenance-logs",
  USER_ACCOUNT_LOGS = "user-account-logs",
  PAYMENT_TRANSACTION_LOGS = "payment-transaction-logs",
}